Improve logs and error handling
