+++ iamctl/pkg/utils/init.go
@@ -1,184 +1,187 @@
-/**
-* Copyright (c) 2023, WSO2 LLC. (https://www.wso2.com) All Rights Reserved.
-*
-* WSO2 LLC. licenses this file to you under the Apache License,
-* Version 2.0 (the "License"); you may not use this file except
-* in compliance with the License.
-* You may obtain a copy of the License at
-*
-* http://www.apache.org/licenses/LICENSE-2.0
-*
-* Unless required by applicable law or agreed to in writing,
-* software distributed under the License is distributed on an
-* "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-* KIND, either express or implied. See the License for the
-* specific language governing permissions and limitations
-* under the License.
+/*
+ * Copyright (c) 2020, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ * WSO2 Inc. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
  */
 
 package utils
 
 import (
-	"crypto/tls"
 	"encoding/json"
 	"fmt"
 	"io/ioutil"
 	"log"
-	"net/http"
-	"net/url"
 	"os"
-	"path/filepath"
-	"strings"
 )
 
-type oAuthResponse struct {
-	AccessToken  string `json:"access_token"`
-	RefreshToken string `json:"refresh_token"`
-	Scope        string `json:"scope"`
-	TokenType    string `json:"token_type"`
-	Expires      int    `json:"expires_in"`
+var dir, _ = os.Getwd()
+var Path = dir + "/iamctl.json"
+var PathSampleSPDetails = dir + "/init.json"
+
+const SCOPE string = "/permission/admin/manage/identity/applicationmgt/update /permission/admin/manage/identity/applicationmgt/create /permission/admin/manage/identity/applicationmgt/view internal_application_mgt_update internal_application_mgt_create internal_application_mgt_view"
+
+const (
+	AppName       = "IAM-CTL"
+	ShortAppDesc  = "Service Provider configuration"
+	LongAPPConfig = "Service Provider configuration"
+)
+
+type SampleSP struct {
+	Server       string `json:"server"`
+	ClientID     string `json:"clientID"`
+	ClientSecret string `json:"clientSecret"`
+	Tenant       string `json:"tenant"`
 }
 
-type ServerConfigs struct {
-	ServerUrl    string `json:"SERVER_URL"`
-	ClientId     string `json:"CLIENT_ID"`
-	ClientSecret string `json:"CLIENT_SECRET"`
-	TenantDomain string `json:"TENANT_DOMAIN"`
-	Username     string `json:"USERNAME"`
-	Password     string `json:"PASSWORD"`
-	Token        string `json:"TOKEN"`
+type ServerDetails struct {
+	Server       string `json:"server"`
+	AccessToken  string `json:"accessToken"`
+	RefreshToken string `json:"refreshToken"`
+}
+type MyJSON struct {
+	Array []ServerDetails
 }
 
-type ToolConfigs struct {
-	KeywordMappings    map[string]interface{} `json:"KEYWORD_MAPPINGS"`
-	ApplicationConfigs map[string]interface{} `json:"APPLICATIONS"`
-}
+func CreateFile() {
 
-type Application struct {
-	Id          string `json:"id"`
-	Name        string `json:"name"`
-	Description string `json:"description"`
-	Self        string `json:"self"`
-}
+	// detect if file exists
+	var _, err = os.Stat(Path)
+	// create file if not exists
+	if os.IsNotExist(err) {
+		var file, err = os.Create(Path)
+		checkError(err)
+		defer file.Close()
 
-type List struct {
-	TotalResults int           `json:"totalResults"`
-	StartIndex   int           `json:"startIndex"`
-	Count        int           `json:"count"`
-	Applications []Application `json:"applications"`
-	Links        []string      `json:"links"`
-}
+		jsonData := &MyJSON{Array: []ServerDetails{}}
+		encodeJson, _ := json.Marshal(jsonData)
 
-var SERVER_CONFIGS ServerConfigs
-var TOOL_CONFIGS ToolConfigs
-
-func LoadConfigs(envConfigPath string) (baseDir string) {
-
-	baseDir = filepath.Dir(filepath.Dir(envConfigPath))
-	serverConfigFile := filepath.Join(envConfigPath, "serverConfig.json")
-	toolConfigFile := filepath.Join(envConfigPath, "toolConfig.json")
-
-	// Load configs from files
-	SERVER_CONFIGS = loadServerConfigsFromFile(serverConfigFile)
-	TOOL_CONFIGS = loadToolConfigsFromFile(toolConfigFile)
-
-	return baseDir
-}
-
-func loadServerConfigsFromFile(configFilePath string) (serverConfigs ServerConfigs) {
-
-	configFile, err := os.Open(configFilePath)
-	if err != nil {
-		log.Fatalln(err.Error())
+		if err != nil {
+			log.Fatalln(err)
+		}
+		err = ioutil.WriteFile(Path, encodeJson, 0644)
+		if err != nil {
+			log.Fatalln(err)
+		}
 	}
-	defer configFile.Close()
-
-	jsonParser := json.NewDecoder(configFile)
-	err = jsonParser.Decode(&serverConfigs)
-	if err != nil {
-		log.Fatalln(err)
-	}
-	fmt.Println("Server configs loaded succesfully from the config file.")
-
-	// Set tenant domain if not defined in the config file
-	if serverConfigs.TenantDomain == "" {
-		log.Println("Tenant domain is not defined in the config file. Using the default tenant domain: carbon.super")
-		serverConfigs.TenantDomain = "carbon.super"
-	}
-
-	// Get access token
-	serverConfigs.Token = getAccessToken(serverConfigs)
-	fmt.Println("Access Token recieved succesfully.")
-
-	return serverConfigs
 }
 
-func loadToolConfigsFromFile(configFilePath string) (toolConfigs ToolConfigs) {
-
-	configFile, err := os.Open(configFilePath)
-	if err != nil {
-		fmt.Println(err.Error())
-		return toolConfigs
-	}
-	defer configFile.Close()
-
-	jsonParser := json.NewDecoder(configFile)
-	jsonParser.Decode(&toolConfigs)
-
-	fmt.Println("Tool configs loaded successfully from the config file.")
-	return toolConfigs
-}
-
-func getAccessToken(config ServerConfigs) string {
+func WriteFiles(server string, token string, refreshToken string) {
 
 	var err error
-	var response oAuthResponse
+	var data MyJSON
+	var msg = new(ServerDetails)
 
-	if config.ServerUrl == "" {
-		log.Fatalln("Server URL is not defined in the config file.")
-	}
-	authUrl := config.ServerUrl + "/oauth2/token"
-
-	body := url.Values{}
-	body.Set("grant_type", "password")
-	body.Set("username", config.Username)
-	body.Set("password", config.Password)
-	body.Set("scope", SCOPE)
-
-	req, err := http.NewRequest("POST", authUrl, strings.NewReader(body.Encode()))
-	if err != nil {
-		log.Fatalln(err)
-	}
-	req.SetBasicAuth(config.ClientId, config.ClientSecret)
-	req.Header.Set("Content-Type", "application/x-www-form-urlencoded")
-	defer req.Body.Close()
-
-	httpClient := &http.Client{
-		Transport: &http.Transport{
-			TLSClientConfig: &tls.Config{
-				InsecureSkipVerify: true,
-			},
-		},
-	}
-	resp, err := httpClient.Do(req)
-	if err != nil {
-		log.Fatalln(err)
-	}
-	defer resp.Body.Close()
-
-	body1, err := ioutil.ReadAll(resp.Body)
+	file, err := ioutil.ReadFile(Path)
 	if err != nil {
 		log.Fatalln(err)
 	}
 
-	if resp.StatusCode != 200 {
-		log.Fatalln("Error in getting access token, response: " + string(body1))
+	err = json.Unmarshal(file, &data)
+	if err != nil {
+		log.Fatalln(err)
 	}
 
-	err2 := json.Unmarshal(body1, &response)
-	if err2 != nil {
-		log.Fatalln(err2)
+	msg.AccessToken = token
+	msg.Server = server
+	msg.RefreshToken = refreshToken
+
+	if len(data.Array) == 0 {
+		data.Array = append(data.Array, *msg)
+	} else {
+		for i := 0; i < len(data.Array); i++ {
+			if data.Array[i].Server == server {
+				data.Array[i].AccessToken = token
+				data.Array[i].RefreshToken = refreshToken
+			} else {
+				data.Array = append(data.Array, *msg)
+			}
+		}
 	}
 
-	return response.AccessToken
+	jsonData, err := json.Marshal(data)
+	if err != nil {
+		log.Fatalln(err)
+	}
+	err = ioutil.WriteFile(Path, jsonData, 0644)
+	if err != nil {
+		log.Fatalln(err)
+	} else {
+		fmt.Println("Authorization is done for : " + server)
+	}
+	checkError(err)
+}
+
+func ReadFile() string {
+
+	var a ServerDetails
+	var data MyJSON
+
+	file, err := ioutil.ReadFile(Path)
+	if err != nil {
+		log.Fatalln(err)
+	}
+
+	err = json.Unmarshal(file, &data)
+	if err != nil {
+		log.Fatalln(err)
+	}
+	//as the single host this worked. For multiple host need to read relevant accessToken according to given server
+	for i := 0; i < len(data.Array); i++ {
+		a = data.Array[i]
+	}
+	return a.AccessToken
+}
+
+func checkError(err error) {
+
+	if err != nil {
+		fmt.Println(err.Error())
+		os.Exit(0)
+	}
+}
+
+func CreateSampleSPFile() {
+
+	// detect if file exists
+	var _, err = os.Stat(PathSampleSPDetails)
+	// create file if not exists
+	if os.IsNotExist(err) {
+		var file, err = os.Create(PathSampleSPDetails)
+		checkError(err)
+		defer file.Close()
+		jsonData := &SampleSP{}
+		encodeJson, _ := json.Marshal(jsonData)
+		if err != nil {
+			log.Fatalln(err)
+		}
+		err = ioutil.WriteFile(PathSampleSPDetails, encodeJson, 0644)
+		if err != nil {
+			log.Fatalln(err)
+		}
+	}
+}
+
+func ReadSPConfig() (string, string, string, string) {
+
+	var data SampleSP
+
+	file, _ := ioutil.ReadFile(PathSampleSPDetails)
+	err := json.Unmarshal(file, &data)
+	if err != nil {
+		log.Fatalln(err)
+	}
+
+	return data.Server, data.ClientID, data.ClientSecret, data.Tenant
 }
